move strtobool to utils omdule
